refactor: Combination field file structure
Moved combination field into combination field array file
